+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/LocaleRepository.php
@@ -19,7 +19,7 @@ class LocaleRepository extends EntityRepository implements LocaleRepositoryInter
     /**
      * {@inheritdoc}
      */
-    public function findBy(array $criteria, array $orderBy = array('code' => 'ASC'), $limit = null, $offset = null)
+    public function findBy(array $criteria, array $orderBy = ['code' => 'ASC'], $limit = null, $offset = null)
     {
         return parent::findBy($criteria, $orderBy, $limit, $offset);
     }
@@ -27,7 +27,7 @@ class LocaleRepository extends EntityRepository implements LocaleRepositoryInter
     /**
      * {@inheritdoc}
      */
-    public function findOneBy(array $criteria, array $orderBy = array('code' => 'ASC'))
+    public function findOneBy(array $criteria, array $orderBy = ['code' => 'ASC'])
     {
         return parent::findOneBy($criteria, $orderBy);
     }
@@ -129,7 +129,7 @@ DQL;
      */
     public function findOneByIdentifier($code)
     {
-        return $this->findOneBy(array('code' => $code));
+        return $this->findOneBy(['code' => $code]);
     }
 
     /**
@@ -137,6 +137,6 @@ DQL;
      */
     public function getIdentifierProperties()
     {
-        return array('code');
+        return ['code'];
     }
 }